Rename `types::random` `types::test_utils`
