+++ beacon_chain/types/src/attestation.rs
@@ -1,7 +1,7 @@
 use super::bls::AggregateSignature;
 use super::ssz::{Decodable, DecodeError, Encodable, SszStream};
 use super::{AttestationData, Bitfield};
-use crate::random::TestRandom;
+use crate::test_utils::TestRandom;
 use rand::RngCore;
 
 #[derive(Debug, Clone, PartialEq)]
@@ -64,7 +64,7 @@ impl<T: RngCore> TestRandom<T> for Attestation {
 mod tests {
     use super::super::ssz::ssz_encode;
     use super::*;
-    use crate::random::TestRandom;
+    use crate::test_utils::TestRandom;
     use rand::{prng::XorShiftRng, SeedableRng};
 
     #[test]
