+++ beacon_chain/types/src/beacon_block.rs
@@ -1,6 +1,6 @@
 use super::ssz::{Decodable, DecodeError, Encodable, SszStream};
 use super::{BeaconBlockBody, Hash256};
-use crate::random::TestRandom;
+use crate::test_utils::TestRandom;
 use bls::AggregateSignature;
 use rand::RngCore;
 
@@ -70,7 +70,7 @@ impl<T: RngCore> TestRandom<T> for BeaconBlock {
 mod tests {
     use super::super::ssz::ssz_encode;
     use super::*;
-    use crate::random::TestRandom;
+    use crate::test_utils::TestRandom;
     use rand::{prng::XorShiftRng, SeedableRng};
 
     #[test]
