+++ beacon_chain/types/src/beacon_block_body.rs
@@ -1,6 +1,6 @@
 use super::ssz::{Decodable, DecodeError, Encodable, SszStream};
 use super::{Attestation, CasperSlashing, Deposit, Exit, ProposerSlashing};
-use crate::random::TestRandom;
+use crate::test_utils::TestRandom;
 use rand::RngCore;
 
 #[derive(Debug, PartialEq, Clone, Default)]
@@ -59,7 +59,7 @@ impl<T: RngCore> TestRandom<T> for BeaconBlockBody {
 mod tests {
     use super::super::ssz::ssz_encode;
     use super::*;
-    use crate::random::TestRandom;
+    use crate::test_utils::TestRandom;
     use rand::{prng::XorShiftRng, SeedableRng};
 
     #[test]
