+++ beacon_chain/types/src/deposit.rs
@@ -1,6 +1,6 @@
 use super::ssz::{Decodable, DecodeError, Encodable, SszStream};
 use super::{DepositData, Hash256};
-use crate::random::TestRandom;
+use crate::test_utils::TestRandom;
 use rand::RngCore;
 
 #[derive(Debug, PartialEq, Clone)]
@@ -49,8 +49,7 @@ impl<T: RngCore> TestRandom<T> for Deposit {
 mod tests {
     use super::super::ssz::ssz_encode;
     use super::*;
-    use crate::random::TestRandom;
-    use rand::{prng::XorShiftRng, SeedableRng};
+    use crate::test_utils::{SeedableRng, TestRandom, XorShiftRng};
 
     #[test]
     pub fn test_ssz_round_trip() {
