+++ beacon_chain/types/src/deposit_input.rs
@@ -1,6 +1,6 @@
 use super::ssz::{decode_ssz_list, Decodable, DecodeError, Encodable, SszStream};
 use super::Hash256;
-use crate::random::TestRandom;
+use crate::test_utils::TestRandom;
 use bls::{PublicKey, Signature};
 use rand::RngCore;
 
@@ -56,7 +56,7 @@ impl<T: RngCore> TestRandom<T> for DepositInput {
 mod tests {
     use super::super::ssz::ssz_encode;
     use super::*;
-    use crate::random::TestRandom;
+    use crate::test_utils::TestRandom;
     use rand::{prng::XorShiftRng, SeedableRng};
 
     #[test]
