+++ beacon_chain/types/src/proposal_signed_data.rs
@@ -1,6 +1,6 @@
 use super::ssz::{Decodable, DecodeError, Encodable, SszStream};
 use super::Hash256;
-use crate::random::TestRandom;
+use crate::test_utils::TestRandom;
 use rand::RngCore;
 
 #[derive(Debug, PartialEq, Clone, Default)]
@@ -49,7 +49,7 @@ impl<T: RngCore> TestRandom<T> for ProposalSignedData {
 mod tests {
     use super::super::ssz::ssz_encode;
     use super::*;
-    use crate::random::TestRandom;
+    use crate::test_utils::TestRandom;
     use rand::{prng::XorShiftRng, SeedableRng};
 
     #[test]
