+++ beacon_chain/types/src/proposer_slashing.rs
@@ -1,6 +1,6 @@
 use super::ssz::{Decodable, DecodeError, Encodable, SszStream};
 use super::ProposalSignedData;
-use crate::random::TestRandom;
+use crate::test_utils::TestRandom;
 use bls::Signature;
 use rand::RngCore;
 
@@ -60,7 +60,7 @@ impl<T: RngCore> TestRandom<T> for ProposerSlashing {
 mod tests {
     use super::super::ssz::ssz_encode;
     use super::*;
-    use crate::random::TestRandom;
+    use crate::test_utils::TestRandom;
     use rand::{prng::XorShiftRng, SeedableRng};
 
     #[test]
