+++ beacon_chain/types/src/slashable_vote_data.rs
@@ -1,6 +1,6 @@
 use super::ssz::{Decodable, DecodeError, Encodable, SszStream};
 use super::AttestationData;
-use crate::random::TestRandom;
+use crate::test_utils::TestRandom;
 use bls::AggregateSignature;
 use rand::RngCore;
 
@@ -55,7 +55,7 @@ impl<T: RngCore> TestRandom<T> for SlashableVoteData {
 mod tests {
     use super::super::ssz::ssz_encode;
     use super::*;
-    use crate::random::TestRandom;
+    use crate::test_utils::TestRandom;
     use rand::{prng::XorShiftRng, SeedableRng};
 
     #[test]
