+++ beacon_chain/types/src/test_utils/aggregate_signature.rs
