+++ beacon_chain/types/src/test_utils/bitfield.rs
