+++ beacon_chain/types/src/test_utils/hash256.rs
