+++ beacon_chain/types/src/test_utils/signature.rs
